+++ stroom-ui/src/api/stroom.ts
@@ -2640,11 +2640,6 @@ export interface SessionListResponse {
   values?: SessionDetails[];
 }
 
-export interface SessionLoginResponse {
-  authenticated?: boolean;
-  redirectUri?: string;
-}
-
 export interface SetAssignedToRequest {
   annotationIdList?: number[];
   assignedTo?: string;
@@ -3291,6 +3286,10 @@ export interface UrlConfig {
   apiKeys?: string;
   changepassword?: string;
   users?: string;
+}
+
+export interface UrlResponse {
+  url?: string;
 }
 
 export interface User {
@@ -4092,24 +4091,6 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
      * No description
      *
      * @tags Authentication
-     * @name Logout
-     * @summary Log a user out of their session
-     * @request GET:/authentication/v1/logout
-     * @secure
-     */
-    logout: (query: { redirect_uri: string }, params: RequestParams = {}) =>
-      this.request<any, boolean>({
-        path: `/authentication/v1/logout`,
-        method: "GET",
-        query: query,
-        secure: true,
-        ...params,
-      }),
-
-    /**
-     * No description
-     *
-     * @tags Authentication
      * @name NeedsPasswordChange
      * @summary Check if a user's password needs changing.
      * @request GET:/authentication/v1/needsPasswordChange
@@ -4212,6 +4193,24 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
         body: data,
         secure: true,
         type: ContentType.Json,
+        ...params,
+      }),
+
+    /**
+     * No description
+     *
+     * @tags Authentication
+     * @name Logout
+     * @summary Log a user out of their session
+     * @request GET:/authentication/v1/noauth/logout
+     * @secure
+     */
+    logout: (query: { redirect_uri: string }, params: RequestParams = {}) =>
+      this.request<any, boolean>({
+        path: `/authentication/v1/noauth/logout`,
+        method: "GET",
+        query: query,
+        secure: true,
         ...params,
       }),
 
@@ -5400,24 +5399,6 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
         body: data,
         secure: true,
         type: ContentType.Json,
-        ...params,
-      }),
-  };
-  gwtStroomSession = {
-    /**
-     * No description
-     *
-     * @tags Stroom Sessions
-     * @name GwtInvalidateStroomSession
-     * @summary Invalidate the current session
-     * @request GET:/gwtStroomSession/v1/invalidate
-     * @secure
-     */
-    gwtInvalidateStroomSession: (params: RequestParams = {}) =>
-      this.request<any, boolean>({
-        path: `/gwtStroomSession/v1/invalidate`,
-        method: "GET",
-        secure: true,
         ...params,
       }),
   };
@@ -7143,15 +7124,16 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
      * No description
      *
      * @tags Sessions
-     * @name LogoutSession
-     * @summary Logs the specified session out of Stroom
-     * @request GET:/session/v1/logout/{sessionId}
+     * @name StroomLogout
+     * @summary Logout of Stroom session
+     * @request GET:/session/v1/logout
      * @secure
      */
-    logoutSession: (sessionId: string, params: RequestParams = {}) =>
-      this.request<any, boolean>({
-        path: `/session/v1/logout/${sessionId}`,
+    stroomLogout: (query: { redirect_uri: string }, params: RequestParams = {}) =>
+      this.request<any, UrlResponse>({
+        path: `/session/v1/logout`,
         method: "GET",
+        query: query,
         secure: true,
         ...params,
       }),
@@ -7160,14 +7142,14 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
      * No description
      *
      * @tags Sessions
-     * @name LoginSession
-     * @summary Checks if the current session is authenticated and redirects to an auth flow if it is not
-     * @request GET:/session/v1/noauth/login
+     * @name ValidateStroomSession
+     * @summary Validate the current session, return a redirect Uri if invalid.
+     * @request GET:/session/v1/noauth/validateSession
      * @secure
      */
-    loginSession: (query?: { redirect_uri?: string }, params: RequestParams = {}) =>
-      this.request<any, SessionLoginResponse>({
-        path: `/session/v1/noauth/login`,
+    validateStroomSession: (query: { redirect_uri: string }, params: RequestParams = {}) =>
+      this.request<any, ValidateSessionResponse>({
+        path: `/session/v1/noauth/validateSession`,
         method: "GET",
         query: query,
         secure: true,
@@ -7780,42 +7762,6 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
         body: data,
         secure: true,
         type: ContentType.Json,
-        ...params,
-      }),
-  };
-  stroomSession = {
-    /**
-     * No description
-     *
-     * @tags Stroom Sessions
-     * @name InvalidateStroomSession
-     * @summary Invalidate the current session
-     * @request GET:/stroomSession/v1/invalidateStroomSession
-     * @secure
-     */
-    invalidateStroomSession: (params: RequestParams = {}) =>
-      this.request<any, boolean>({
-        path: `/stroomSession/v1/invalidateStroomSession`,
-        method: "GET",
-        secure: true,
-        ...params,
-      }),
-
-    /**
-     * No description
-     *
-     * @tags Stroom Sessions
-     * @name ValidateStroomSession
-     * @summary Validate the current session, return a redirect Uri if invalid.
-     * @request GET:/stroomSession/v1/noauth/validateSession
-     * @secure
-     */
-    validateStroomSession: (query: { redirect_uri: string }, params: RequestParams = {}) =>
-      this.request<any, ValidateSessionResponse>({
-        path: `/stroomSession/v1/noauth/validateSession`,
-        method: "GET",
-        query: query,
-        secure: true,
         ...params,
       }),
   };