+++ stroom-ui/src/startup/Authorisation/index.ts
@@ -1,9 +1,0 @@
-import AuthorisationContext from "./AuthorisationContext";
-import AuthorisationContextProvider from "./AuthorisationContextProvider";
-import useAuthorisationContext from "./useAuthorisationContext";
-
-export {
-  AuthorisationContext,
-  AuthorisationContextProvider,
-  useAuthorisationContext,
-};